refactor and update `numberRow` between swiss and arenas to show on hover the number of games involved when displaying a percent
