+++ ui/swiss/src/view/util.ts
@@ -1,6 +1,5 @@
 import { h } from 'snabbdom';
 import type { BasePlayer } from '../interfaces';
-import { numberFormat } from 'lib/i18n';
 import { fullName, userLine, userRating } from 'lib/view/userLink';
 
 export function player(p: BasePlayer, asLink: boolean, withRating: boolean) {
@@ -16,22 +15,4 @@ export function player(p: BasePlayer, asLink: boolean, withRating: boolean) {
       withRating ? h('span.rating', userRating({ ...p, brackets: false })) : null,
     ],
   );
-}
-
-export const ratio2percent = (r: number) => Math.round(100 * r) + '%';
-
-export function numberRow(name: string, value: any, typ?: string) {
-  return h('tr', [
-    h('th', name),
-    h(
-      'td',
-      typ === 'raw'
-        ? value
-        : typ === 'percent'
-          ? value[1] > 0
-            ? ratio2percent(value[0] / value[1])
-            : 0
-          : numberFormat(value),
-    ),
-  ]);
 }
