+++ ui/tournament/src/view/finished.ts
@@ -8,7 +8,7 @@ import { teamStanding } from './battle';
 import header from './header';
 import playerInfo from './playerInfo';
 import teamInfo from './teamInfo';
-import { numberRow } from './util';
+import { numberRow } from 'lib/view/util';
 import { type MaybeVNodes } from 'lib/snabbdom';
 import { once } from 'lib/storage';
