+++ ui/tournament/src/view/playerInfo.ts
@@ -1,8 +1,9 @@
 import * as licon from 'lib/licon';
 import { spinnerVdom as spinner } from 'lib/view/controls';
 import { type VNode, bind, dataIcon, hl } from 'lib/snabbdom';
-import { numberRow, player as renderPlayer } from './util';
+import { player as renderPlayer } from './util';
 import { fullName } from 'lib/view/userLink';
+import { numberRow } from 'lib/view/util';
 import { teamName } from './battle';
 import { status } from 'lib/game/game';
 import type TournamentController from '../ctrl';
